# dist/routes/swagger.json
{
    "components": {
        "examples": {},
        "headers": {},
        "parameters": {},
        "requestBodies": {},
        "responses": {},
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "anyOf": [
                            {},
                            {
                                "items": {},
                                "type": "array"
                            }
                        ]
                    },
                    "total": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "message",
                    "data"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IStudent": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "course": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "course"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "node-ts",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
        "/student/list": {
            "get": {
                "operationId": "StudentList",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Student"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/student/{studentId}/detail": {
            "get": {
                "operationId": "StudentDetail",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Student"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "studentId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/student/create-student": {
            "post": {
                "operationId": "AddStudent",
                "responses": {
                    "201": {
                        "description": "Student Registered successfully"
                    }
                },
                "tags": [
                    "Student"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IStudent"
                            }
                        }
                    }
                }
            }
        },
        "/student/{studentId}/delete": {
            "delete": {
                "operationId": "DeleteStudent",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Student"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "studentId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/student/{studentId}/update": {
            "put": {
                "operationId": "UpdateStudentDetail",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "Student"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "studentId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IStudent"
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/v1"
        }
    ]
}